+++ src/main/java/com/supermartijn642/trashcans/filter/IFilterManager.java
@@ -1,8 +1,7 @@
 package com.supermartijn642.trashcans.filter;
 
-import net.minecraft.core.HolderLookup;
-import net.minecraft.nbt.Tag;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.storage.ValueInput;
 
 /**
  * Created 12/19/2020 by SuperMartijn642
@@ -11,6 +10,6 @@ public interface IFilterManager {
 
     ItemFilter createFilter(ItemStack stack);
 
-    ItemFilter readFilter(Tag tag, HolderLookup.Provider provider);
+    ItemFilter readFilter(ValueInput input);
 
 }