+++ changelog.md
@@ -1,2 +1,2 @@
 ### Trash Cans 1.0.18
-- Initial release of Trash Cans for Forge 1.21.5
+- Initial release of Trash Cans for Minecraft 1.21.6 & 1.21.7